setupPageIndicators() 가 페이지 담당 코드 구현 및 코드 수정
