+++ Health/Presentation/Onboarding/OnboardingViews/viewcontrollers/HeightViewController.swift
@@ -1,5 +1,5 @@
 //
-//  WeightViewController.swift
+//  HeightViewController.swift
 //  Health
 //
 //  Created by 권도현 on 8/4/25.
@@ -7,13 +7,13 @@
 
 import UIKit
 
-class WeightViewController: CoreViewController {
+class HeightViewController: CoreViewController {
+
+    @IBOutlet weak var heightInputField: UITextField!
     
-    @IBOutlet weak var weightInputField: UITextField!
-    
-    private let kgLabel: UILabel = {
+    private let cmLabel: UILabel = {
         let label = UILabel()
-        label.text = "kg"
+        label.text = "cm"
         label.textColor = .accent
         label.font = UIFont.systemFont(ofSize: 16, weight: .medium)
         label.translatesAutoresizingMaskIntoConstraints = false
@@ -30,26 +30,18 @@ class WeightViewController: CoreViewController {
         return button
     }()
     
-    private let pageIndicatorStack: UIStackView = {
-        let stack = UIStackView()
-        stack.axis = .horizontal
-        stack.spacing = 6
-        stack.distribution = .fillEqually
-        return stack
-    }()
+    private let progressIndicatorView = ProgressIndicatorView(totalPages: 4)
 
     private var continueButtonBottomConstraint: NSLayoutConstraint?
 
-    override func initVM() { }
+    override func initVM() {}
 
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        weightInputField.delegate = self
-        weightInputField.keyboardType = .numberPad
-        weightInputField.addTarget(self, action: #selector(textFieldDidChange), for: .editingChanged)
-       
-        continueButton.addTarget(self, action: #selector(didTapContinue), for: .touchUpInside)
+        heightInputField.delegate = self
+        heightInputField.keyboardType = .numberPad
+        heightInputField.addTarget(self, action: #selector(textFieldDidChange), for: .editingChanged)
         
         registerForKeyboardNotifications()
         setupTapGestureToDismissKeyboard()
@@ -59,23 +51,23 @@ class WeightViewController: CoreViewController {
 
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
-        pageIndicatorStack.isHidden = false
-    }
-    
-    override func viewWillDisappear(_ animated: Bool) {
-        super.viewWillDisappear(animated)
-        pageIndicatorStack.isHidden = true
+        progressIndicatorView.isHidden = false
     }
 
+    override func viewWillDisappear(_ animated: Bool) {
+        super.viewWillDisappear(animated)
+        progressIndicatorView.isHidden = true
+    }
+    
     override func setupHierarchy() {
-        [continueButton, pageIndicatorStack, kgLabel].forEach {
+        [continueButton, progressIndicatorView, cmLabel].forEach {
             $0.translatesAutoresizingMaskIntoConstraints = false
             view.addSubview($0)
         }
     }
 
     override func setupAttribute() {
-        setupPageIndicators(progress: 0.5)
+        progressIndicatorView.updateProgress(to: 0.625)
     }
 
     override func setupConstraints() {
@@ -87,77 +79,30 @@ class WeightViewController: CoreViewController {
             continueButton.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -20),
             continueButton.heightAnchor.constraint(equalToConstant: 48),
             
-            pageIndicatorStack.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: -24),
-            pageIndicatorStack.centerXAnchor.constraint(equalTo: view.centerXAnchor),
-            pageIndicatorStack.heightAnchor.constraint(equalToConstant: 4),
-            pageIndicatorStack.widthAnchor.constraint(equalToConstant: 320),
+            progressIndicatorView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: -24),
+            progressIndicatorView.centerXAnchor.constraint(equalTo: view.centerXAnchor),
+            progressIndicatorView.heightAnchor.constraint(equalToConstant: 4),
+            progressIndicatorView.widthAnchor.constraint(equalToConstant: 320),
             
-            kgLabel.leadingAnchor.constraint(equalTo: weightInputField.trailingAnchor, constant: 8),
-            kgLabel.centerYAnchor.constraint(equalTo: weightInputField.centerYAnchor)
+            cmLabel.leadingAnchor.constraint(equalTo: heightInputField.trailingAnchor, constant: 8),
+            cmLabel.centerYAnchor.constraint(equalTo: heightInputField.centerYAnchor)
         ])
     }
-
-    private func setupPageIndicators(progress: CGFloat) {
-        pageIndicatorStack.arrangedSubviews.forEach { $0.removeFromSuperview() }
-
-        let totalPages = 4
-        let clampedProgress = max(0, min(progress, 1))
-        let totalProgress = CGFloat(totalPages) * clampedProgress
-
-        for i in 0..<totalPages {
-            let containerView = UIView()
-            containerView.backgroundColor = .buttonBackground
-            containerView.layer.cornerRadius = 2
-            containerView.clipsToBounds = true
-
-            let progressBar = UIView()
-            progressBar.backgroundColor = .accent
-            progressBar.translatesAutoresizingMaskIntoConstraints = false
-
-            containerView.addSubview(progressBar)
-
-            let fillRatio: CGFloat
-            if totalProgress > CGFloat(i + 1) {
-                fillRatio = 1.0
-            } else if totalProgress > CGFloat(i) {
-                fillRatio = totalProgress - CGFloat(i)
-            } else {
-                fillRatio = 0.0
-            }
-
-            NSLayoutConstraint.activate([
-                progressBar.leadingAnchor.constraint(equalTo: containerView.leadingAnchor),
-                progressBar.topAnchor.constraint(equalTo: containerView.topAnchor),
-                progressBar.bottomAnchor.constraint(equalTo: containerView.bottomAnchor),
-                progressBar.widthAnchor.constraint(equalTo: containerView.widthAnchor, multiplier: fillRatio)
-            ])
-
-            pageIndicatorStack.addArrangedSubview(containerView)
-
-            containerView.translatesAutoresizingMaskIntoConstraints = false
-            containerView.heightAnchor.constraint(equalToConstant: 4).isActive = true
-        }
-    }
-
+    
     @objc private func textFieldDidChange(_ textField: UITextField) {
         validateInput()
     }
 
     private func validateInput() {
-        guard let text = weightInputField.text, let weight = Int(text), (40...200).contains(weight) else {
+        guard let text = heightInputField.text, let weight = Int(text), (100...250).contains(weight) else {
             continueButton.isEnabled = false
             continueButton.backgroundColor = .buttonBackground
-            weightInputField.textColor = .label
+            heightInputField.textColor = .label
             return
         }
         continueButton.isEnabled = true
         continueButton.backgroundColor = .accent
-        weightInputField.textColor = .accent
-    }
-
-    @objc private func didTapContinue() {
-        guard continueButton.isEnabled else { return }
-        performSegue(withIdentifier: "goToHeightInfo", sender: nil)
+        heightInputField.textColor = .accent
     }
 
     private func registerForKeyboardNotifications() {
@@ -195,8 +140,9 @@ class WeightViewController: CoreViewController {
     }
 }
 
-extension WeightViewController: UITextFieldDelegate {
+extension HeightViewController: UITextFieldDelegate {
     func textField(_ textField: UITextField, shouldChangeCharactersIn range: NSRange, replacementString string: String) -> Bool {
+        
         let allowedCharacters = CharacterSet.decimalDigits
         if string.rangeOfCharacter(from: allowedCharacters.inverted) != nil {
             return false
