+++ Health/Presentation/Onboarding/OnboardingViews/viewcontrollers/SelectAgeViewController.swift
@@ -1,24 +1,8 @@
-//
-//  HeightViewController.swift
-//  Health
-//
-//  Created by 권도현 on 8/4/25.
-//
-
 import UIKit
 
-class HeightViewController: CoreViewController {
-
-    @IBOutlet weak var heightInputField: UITextField!
+class SelectAgeViewController: CoreViewController {
     
-    private let cmLabel: UILabel = {
-        let label = UILabel()
-        label.text = "cm"
-        label.textColor = .accent
-        label.font = UIFont.systemFont(ofSize: 16, weight: .medium)
-        label.translatesAutoresizingMaskIntoConstraints = false
-        return label
-    }()
+    @IBOutlet weak var ageInputField: UITextField!
     
     private let continueButton: UIButton = {
         let button = UIButton(type: .system)
@@ -29,25 +13,21 @@ class HeightViewController: CoreViewController {
         button.isEnabled = false
         return button
     }()
-    
-    private let pageIndicatorStack: UIStackView = {
-        let stack = UIStackView()
-        stack.axis = .horizontal
-        stack.spacing = 6
-        stack.distribution = .fillEqually
-        return stack
-    }()
 
     private var continueButtonBottomConstraint: NSLayoutConstraint?
 
-    override func initVM() {}
+    private let progressIndicatorView = ProgressIndicatorView(totalPages: 4)
+    
+    override func initVM() { }
 
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        heightInputField.delegate = self
-        heightInputField.keyboardType = .numberPad
-        heightInputField.addTarget(self, action: #selector(textFieldDidChange), for: .editingChanged)
+        ageInputField.delegate = self
+        ageInputField.keyboardType = .numberPad
+        ageInputField.addTarget(self, action: #selector(textFieldDidChange), for: .editingChanged)
+        
+        continueButton.addTarget(self, action: #selector(didTapContinue), for: .touchUpInside)
         
         registerForKeyboardNotifications()
         setupTapGestureToDismissKeyboard()
@@ -57,23 +37,23 @@ class HeightViewController: CoreViewController {
 
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
-        pageIndicatorStack.isHidden = false
-    }
-
-    override func viewWillDisappear(_ animated: Bool) {
-        super.viewWillDisappear(animated)
-        pageIndicatorStack.isHidden = true
+        progressIndicatorView.isHidden = false
     }
     
+    override func viewWillDisappear(_ animated: Bool) {
+        super.viewWillDisappear(animated)
+        progressIndicatorView.isHidden = true
+    }
+
     override func setupHierarchy() {
-        [continueButton, pageIndicatorStack, cmLabel].forEach {
+        [continueButton, progressIndicatorView].forEach {
             $0.translatesAutoresizingMaskIntoConstraints = false
             view.addSubview($0)
         }
     }
 
     override func setupAttribute() {
-        setupPageIndicators(progress: 0.625)
+        progressIndicatorView.updateProgress(to: 0.375)
     }
 
     override func setupConstraints() {
@@ -85,71 +65,32 @@ class HeightViewController: CoreViewController {
             continueButton.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -20),
             continueButton.heightAnchor.constraint(equalToConstant: 48),
             
-            pageIndicatorStack.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: -24),
-            pageIndicatorStack.centerXAnchor.constraint(equalTo: view.centerXAnchor),
-            pageIndicatorStack.heightAnchor.constraint(equalToConstant: 4),
-            pageIndicatorStack.widthAnchor.constraint(equalToConstant: 320),
-            
-            cmLabel.leadingAnchor.constraint(equalTo: heightInputField.trailingAnchor, constant: 8),
-            cmLabel.centerYAnchor.constraint(equalTo: heightInputField.centerYAnchor)
+            progressIndicatorView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: -24),
+            progressIndicatorView.centerXAnchor.constraint(equalTo: view.centerXAnchor),
+            progressIndicatorView.heightAnchor.constraint(equalToConstant: 4),
+            progressIndicatorView.widthAnchor.constraint(equalToConstant: 320)
         ])
     }
 
-    private func setupPageIndicators(progress: CGFloat) {
-        pageIndicatorStack.arrangedSubviews.forEach { $0.removeFromSuperview() }
-
-        let totalPages = 4
-        let clampedProgress = max(0, min(progress, 1))
-        let totalProgress = CGFloat(totalPages) * clampedProgress
-
-        for i in 0..<totalPages {
-            let containerView = UIView()
-            containerView.backgroundColor = .buttonBackground
-            containerView.layer.cornerRadius = 2
-            containerView.clipsToBounds = true
-
-            let progressBar = UIView()
-            progressBar.backgroundColor = .accent
-            progressBar.translatesAutoresizingMaskIntoConstraints = false
-
-            containerView.addSubview(progressBar)
-
-            let fillRatio: CGFloat
-            if totalProgress > CGFloat(i + 1) {
-                fillRatio = 1.0
-            } else if totalProgress > CGFloat(i) {
-                fillRatio = totalProgress - CGFloat(i)
-            } else {
-                fillRatio = 0.0
-            }
-
-            NSLayoutConstraint.activate([
-                progressBar.leadingAnchor.constraint(equalTo: containerView.leadingAnchor),
-                progressBar.topAnchor.constraint(equalTo: containerView.topAnchor),
-                progressBar.bottomAnchor.constraint(equalTo: containerView.bottomAnchor),
-                progressBar.widthAnchor.constraint(equalTo: containerView.widthAnchor, multiplier: fillRatio)
-            ])
-
-            pageIndicatorStack.addArrangedSubview(containerView)
-
-            containerView.translatesAutoresizingMaskIntoConstraints = false
-            containerView.heightAnchor.constraint(equalToConstant: 4).isActive = true
-        }
-    }
     @objc private func textFieldDidChange(_ textField: UITextField) {
         validateInput()
     }
 
     private func validateInput() {
-        guard let text = heightInputField.text, let weight = Int(text), (100...250).contains(weight) else {
+        guard let text = ageInputField.text, let year = Int(text), (1900...2025).contains(year) else {
             continueButton.isEnabled = false
             continueButton.backgroundColor = .buttonBackground
-            heightInputField.textColor = .label
+            ageInputField.textColor = .label // 기본색
             return
         }
         continueButton.isEnabled = true
         continueButton.backgroundColor = .accent
-        heightInputField.textColor = .accent
+        ageInputField.textColor = .accent
+    }
+    
+    @objc private func didTapContinue() {
+        guard continueButton.isEnabled else { return }
+        performSegue(withIdentifier: "goToWeightInfo", sender: nil)
     }
 
     private func registerForKeyboardNotifications() {
@@ -187,9 +128,8 @@ class HeightViewController: CoreViewController {
     }
 }
 
-extension HeightViewController: UITextFieldDelegate {
+extension SelectAgeViewController: UITextFieldDelegate {
     func textField(_ textField: UITextField, shouldChangeCharactersIn range: NSRange, replacementString string: String) -> Bool {
-        
         let allowedCharacters = CharacterSet.decimalDigits
         if string.rangeOfCharacter(from: allowedCharacters.inverted) != nil {
             return false
@@ -197,7 +137,7 @@ extension HeightViewController: UITextFieldDelegate {
 
         let currentText = textField.text ?? ""
         let prospectiveText = (currentText as NSString).replacingCharacters(in: range, with: string)
-        return prospectiveText.count <= 3
+        return prospectiveText.count <= 4
     }
 
     func textFieldShouldReturn(_ textField: UITextField) -> Bool {
